Refactor environment variable handling by consolidating getEnvVar function into utils.ts. Update import paths across deployment scripts to utilize the new location. Introduce proposeDiamondCut function in shared directory for diamond cut proposals, enhancing modularity and reducing cyclic dependencies.
